♻️ Split common code to AbstractPackagePool
